Working with new data
+ use the new life_expectancy_Cohort_ages.csv 
+ notice that "WORLD" is now "World"
+ notice that there are only male or female ("both" category is deleted)
+ The matrix now only compares the Non-zero numbers and does it in a "smart" way.
This bit of code: "(np.amax(max(np.where(life_exp_[:,1] == 0))))" says -- in the column of the matrix, find the max index number of the zero and subsequently use this to subset the rows of the matrix column. This ensures that the interpolation is operating on two columns of even length.